pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "aetherium-qcom"
version = "0.1.0"
authors = [
  { name="Yaron Koresh", email="aharonkoresh1@gmail.com" },
]
description = "Aetherium Q-Com is a decentralized, untraceable, and unblockable communication platform engineered for truly private, censorship-resistant messaging."
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE" }
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Topic :: Communications :: Chat",
    "Topic :: Security :: Cryptography",
]
dependencies = [
    "kademlia",
    "cryptography",
    "pqcrypto",
    "numpy",
    "pillow",
    "pydub",
    "moviepy",
    "pyside6",
    "olefile",
    "defusedxml",
    "pillow-heif"
]

[project.urls]
"Homepage" = "https://github.com/YaronKoresh/aetherium-qcom"
"Bug Tracker" = "https://github.com/YaronKoresh/aetherium-qcom/issues"